requirements.txt
PySide6>=6.5
Pillow>=9.5
pillow-heif>=0.13
